Out of process modules for Azure IoT Gateway.
Run modules in a separate process from the gateway.

Feature consists of an "outprocess" loader and module, which acts as the proxy
for the remote module in the gateway. Conversely, there is a Proxy_Gateway
library for the external process, which acts as the gateway proxy for the
remote module. The outprocess module and proxy gateway communicate through a
control channel, which must be given to each process.

The gateway assumes no control over the external process.  The proxy gateway
library allows a module to attach and detach from the gateway, in case the
gateway module's lifetime is different than the external process's lifetime.

+++ core/tests/broker_ut/broker_ut.cpp
@@ -547,7 +547,7 @@ public:
         int send_length;
         if (len == NN_MSG)
         {
-            send_length = nn_current_msg_size;
+            send_length = (int)nn_current_msg_size;
             free(*(void**)buf); // send is supposed to free auto created buffer on success
         }
         else
